fix: Move async custom logger tests to test_litellm directory
Per reviewer feedback, moved the new async callback tests from
tests/local_testing/test_custom_logger.py to a new file
tests/test_litellm/test_async_custom_logger.py to follow
the project's test directory structure conventions.

Tests are passing with the new location.
